reorganize project folder and files structure
