--- src/components/ModalMemo.tsx
@@ -1,4 +1,0 @@
-import React from "react";
-import Modal from "./Modal";
-
-export const ModalMemo = React.memo(Modal);
